Refactoring App Services. Added App Config entries for VMs and App service backend services. Need to test deployment.
