+++ app/ADE.Services/src/ADE.ServiceBase/Program.cs
@@ -1,0 +1,10 @@
+namespace ADE.ServiceBase
+{
+    public class Program
+    {
+        public static void Main(string[] args)
+        {
+            // Exit - this is a dummy method
+        }
+    }
+}
