+++ deployments/azure_app_services/azure_app_services.parameters.sample.json
@@ -8,9 +8,6 @@
         "aliasRegion": {
             "value": "@@aliasRegion@@"
         },
-        "rootDomainName": {
-            "value": "@@rootDomainName@@"
-        },
         "appServicePlanResourceGroupName": {
             "value": "@@appServicePlanResourceGroupName@@"
         },
