+++ deployments/azure_app_services/azure_app_services_adeapp_app_config.bicep
@@ -1,0 +1,10 @@
+param aliasRegion string
+param appConfigName string
+param backendServices array
+
+resource appConfigKeyVirtualMachineBackingServiceConnectionStrings 'Microsoft.AppConfiguration/configurationStores/keyValues@2020-07-01-preview' = [for backendService in backendServices: {
+  name: '${appConfigName}/ADE:${backendService.name}Uri$appservices'
+  properties: {
+    value: 'https://${replace('app-ade-${aliasRegion}-ade-${backendService.name}', '-', '')}.azurewebsites.net/'
+  }
+}]
