+++ deployments/azure_virtual_machines/adeappinstall.sh
@@ -90,11 +90,12 @@ sudo docker pull $ACR_SERVER.azurecr.io/ade-apigateway:latest
 
 echo "Starting Frontend ADE Service"
 
-sudo docker run --name "ade-frontend" -d --restart unless-stopped -p 80:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" $ACR_SERVER.azurecr.io/ade-frontend:latest
+sudo docker run --name "ade-frontend" -d --restart unless-stopped -p 80:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" -e ADE__ENVIRONMENT="virtualmachines" $ACR_SERVER.azurecr.io/ade-frontend:latest
 
 # external api gateway - note, we override the connection info to our local docker instances
 sudo docker run --name "ade-apigateway" -d --restart unless-stopped -p 8080:80 \\
     -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" \\
+    -e ADE__ENVIRONMENT="virtualmachines" \\
     -e ADE__DATAINGESTORSERVICEURI="http://$ADE_BACKEND_IPADDRESS:5000" \\
     -e ADE__DATAREPORTERSERVICEURI="http://$ADE_BACKEND_IPADDRESS:5001" \\
     -e ADE__EVENTINGESTORSERVICEURI="http://$ADE_BACKEND_IPADDRESS:5002" \\
@@ -116,10 +117,10 @@ sudo docker pull $ACR_SERVER.azurecr.io/ade-eventingestorservice:latest
 echo "Starting Backend ADE Services"
 
 # local docker network services
-sudo docker run --name "ade-dataingestorservice" -d --restart unless-stopped -p 5000:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" $ACR_SERVER.azurecr.io/ade-dataingestorservice:latest
-sudo docker run --name "ade-datareporterservice" -d --restart unless-stopped -p 5001:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" $ACR_SERVER.azurecr.io/ade-datareporterservice:latest
-sudo docker run --name "ade-userservice" -d --restart unless-stopped -p 5002:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" $ACR_SERVER.azurecr.io/ade-userservice:latest
-sudo docker run --name "ade-eventingestorservice" -d --restart unless-stopped -p 5003:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" $ACR_SERVER.azurecr.io/ade-eventingestorservice:latest
+sudo docker run --name "ade-dataingestorservice" -d --restart unless-stopped -p 5000:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" -e ADE__ENVIRONMENT="virtualmachines" $ACR_SERVER.azurecr.io/ade-dataingestorservice:latest
+sudo docker run --name "ade-datareporterservice" -d --restart unless-stopped -p 5001:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" -e ADE__ENVIRONMENT="virtualmachines" $ACR_SERVER.azurecr.io/ade-datareporterservice:latest
+sudo docker run --name "ade-userservice" -d --restart unless-stopped -p 5002:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" -e ADE__ENVIRONMENT="virtualmachines" $ACR_SERVER.azurecr.io/ade-userservice:latest
+sudo docker run --name "ade-eventingestorservice" -d --restart unless-stopped -p 5003:80 -e CONNECTIONSTRINGS__APPCONFIG="$APPCONFIG_CONNECTIONSTRING" -e ADE__ENVIRONMENT="virtualmachines" $ACR_SERVER.azurecr.io/ade-eventingestorservice:latest
 EOF
 fi
