+++ deployments/azure_virtual_machines/azure_virtual_machines_ntier_app_config.bicep
@@ -1,0 +1,9 @@
+param appConfigName string
+param backendServices array
+
+resource appConfigKeyVirtualMachineBackingServiceConnectionStrings 'Microsoft.AppConfiguration/configurationStores/keyValues@2020-07-01-preview' = [for backendService in backendServices: {
+  name: '${appConfigName}/ADE:${backendService.name}Uri$virtualmachines'
+  properties: {
+    value: 'http://localhost:${backendService.port}/'
+  }
+}]
